New controller for downloading torrents
Moved the code for downloading the torrent (server-side) to its own
controller. Also makes the request handling easier.
