+++ templates/upload_template.html
@@ -20,6 +20,7 @@ Server is now downloading and will be seeding: {{title}}.<br>You may go back to 
 <script src="https://cdn.jsdelivr.net/webtorrent/latest/webtorrent.min.js"></script>
 <script>
 var client = new WebTorrent()
+var xhttp = new XMLHttpRequest()
 
 client.on('error', (err) => {
 	console.error('ERROR: ' + err.message)
@@ -37,25 +38,24 @@ function handlesubmit (submit) {
 
 //Using the data from the form, seed the torrent, or print error
 function startseed (formData) {
-  console.log('1:' + formData)
   torrentfile = formData.getAll('file')
   client.seed(torrentfile, { announceList: [['ws://localhost:8080']] })
-  console.log('2:' + formData)
-
-  //Once the torrent starts seeding, inform the server to download the torrent
   client.on('torrent', function (torrent) {
-    //Send a notification to the server to download this file.
-    //TODO: find a better way.
-    console.log('3:' + formData)
-    var xhttp = new XMLHttpRequest()
-    xhttp.open('POST', 'http://localhost/upload', true)
-    var sendData = new FormData()
-    sendData.append('title', formData.get('title'))
-    sendData.append('desc', formData.get('desc'))
-    sendData.append('infoHash', torrent.infoHash)
-    xhttp.send(sendData)
-    console.log('Seeding file')
+    torrent.on('done', function() {
+      xhttp.open('POST', 'http://localhost/upload', true)
+      xhttp.onreadystatechange = handleresponse
+      var sendData = new FormData()
+      sendData.append('title', formData.get('title'))
+      sendData.append('desc', formData.get('desc'))
+      sendData.append('infoHash', torrent.infoHash)
+      xhttp.send(sendData)
+      console.log('Seeding file')
+    })
   })
+  function handleresponse() {
+    console.log(xhttp.responseText)
+  }
 }
+
 
 </script>
